a user can only delete and update his own recipies only

+++ Recipes/task/src/recipes/RecipeController.java
@@ -6,9 +6,15 @@ import com.google.gson.JsonObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.*;
 import org.springframework.web.server.ResponseStatusException;
+import recipes.users.User;
+import recipes.users.UserDetailsImpl;
+import recipes.users.UserService;
 
 
 import javax.validation.Valid;
@@ -22,17 +28,22 @@ public class RecipeController {
 
     private final RecipeService recipeService;
 
-
+    private final UserService userService;
 
     @Autowired
-    public RecipeController(RecipeService recipeService){
+    public RecipeController(RecipeService recipeService, UserService userService){
         this.recipeService = recipeService;
+        this.userService = userService;
     }
 
     @PostMapping("/api/recipe/new")
-    public ResponseEntity<String> addRecipe(@Valid @RequestBody Recipe recipe) {
+    public ResponseEntity<String> addRecipe(@Valid @RequestBody Recipe recipe, @AuthenticationPrincipal UserDetailsImpl user) {
+
+        User currentUser = userService.findByEmail(user.getUsername());
+        currentUser.assignRecipe(recipe);
 
         Recipe newRecipe = recipeService.save(recipe);
+
 
         JsonObject json = new JsonObject();
         json.addProperty("id", newRecipe.getId());
@@ -47,14 +58,27 @@ public class RecipeController {
     }
 
     @DeleteMapping("/api/recipe/{id}")
-    public ResponseEntity<String> deleteRecipe(@PathVariable Long id) {
+    public ResponseEntity<String> deleteRecipe(@PathVariable Long id, @AuthenticationPrincipal UserDetailsImpl user) {
+
+        User currentUser = userService.findByEmail(user.getUsername());
+        if(!currentUser.hasRecipe(recipeService.findRecipeById(id))) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        currentUser.removeRecipe(recipeService.findRecipeById(id));
         recipeService.deleteRecipeById(id);
 
         return ResponseEntity.noContent().build();
     }
 
     @PutMapping("/api/recipe/{id}")
-    public ResponseEntity<Recipe> updateRecipe(@PathVariable Long id, @Valid @RequestBody Recipe updatedRecipe){
+    public ResponseEntity<Recipe> updateRecipe(@PathVariable Long id, @Valid @RequestBody Recipe updatedRecipe, @AuthenticationPrincipal UserDetailsImpl user){
+
+        User currentUser = userService.findByEmail(user.getUsername());
+        if(!currentUser.hasRecipe(recipeService.findRecipeById(id))) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         recipeService.updateRecipe(id,updatedRecipe);
         return ResponseEntity.noContent().build();
     }
